+++ config/config-gofer.hcl
@@ -209,7 +209,6 @@ gofer {
       min_values = 3
       origin "bitstamp" { query = "ADA/BTC" }
       origin "bitfinex_simple" { query = "ADA/BTC" }
-      origin "coinbase" { query = "ADA/BTC" }
       origin "kraken" { query = "ADA/BTC" }
       origin "okx" { query = "ADA/BTC" }
       origin "upbit" { query = "ADA/BTC" }
